[flow] refactor types imports and fix some issues
