+++ app/javascript/src/Policies/actions/OriginalPolicyChain.jsx
@@ -1,6 +1,6 @@
 // @flow
 
-import type { ChainPolicy } from 'Policies/types/Policies'
+import type { ChainPolicy } from 'Policies/types'
 
 export type SetOriginalPolicyChainAction = { type: 'SET_ORIGINAL_POLICY_CHAIN', payload: Array<ChainPolicy> }
 export function setOriginalPolicyChain (payload: Array<ChainPolicy>): SetOriginalPolicyChainAction {
