+++ app/javascript/src/Policies/actions/PolicyConfig.jsx
@@ -1,6 +1,6 @@
 // @flow
 
-import type { ChainPolicy } from 'Policies/types/Policies'
+import type { ChainPolicy } from 'Policies/types'
 
 export type UpdatePolicyConfigAction = { type: 'UPDATE_POLICY_CONFIG', policyConfig: ChainPolicy }
 export function updatePolicyConfig (policyConfig: ChainPolicy): UpdatePolicyConfigAction {
