+++ app/javascript/src/Policies/actions/index.jsx
@@ -15,10 +15,7 @@ import {
 import { updatePolicyConfig } from 'Policies/actions/PolicyConfig'
 
 import type { UIComponent } from 'Policies/actions/UISettings'
-import type { SortPolicyChainAction } from 'Policies/actions/PolicyChain'
-import type { UpdatePolicyConfigAction } from 'Policies/actions/PolicyConfig'
-import type { Dispatch, ThunkAction } from 'Policies/types/index'
-import type { RawRegistry, RegistryPolicy, ChainPolicy, StoredChainPolicy } from 'Policies/types/Policies'
+import type { Dispatch, RawRegistry, RegistryPolicy, ChainPolicy, StoredChainPolicy, IPoliciesActions, ThunkAction } from 'Policies/types'
 
 const chainComponent: UIComponent = 'chain'
 const registryComponent: UIComponent = 'registry'
@@ -91,19 +88,6 @@ function closePolicyConfig (): ThunkAction {
   return function (dispatch: Dispatch) {
     dispatch(uiComponentTransition({hide: policyConfigComponent, show: chainComponent}))
   }
-}
-
-export interface IPoliciesActions {
-  openPolicyRegistry: () => ThunkAction,
-  editPolicy: ChainPolicy => ThunkAction,
-  sortPolicyChain: Array<ChainPolicy> => SortPolicyChainAction,
-  submitPolicyConfig: (ChainPolicy) => ThunkAction,
-  removePolicyFromChain: (ChainPolicy) => ThunkAction,
-  closePolicyConfig: () => ThunkAction,
-  addPolicy: RegistryPolicy => ThunkAction,
-  closePolicyRegistry: () => ThunkAction,
-  populatePolicies: (serviceId: string, chain?: Array<StoredChainPolicy>, registry?: RawRegistry) => ThunkAction,
-  updatePolicyConfig: (ChainPolicy) => UpdatePolicyConfigAction
 }
 
 export const actions: IPoliciesActions = {
