+++ app/javascript/src/Policies/components/PoliciesForm.jsx
@@ -5,8 +5,7 @@ import Form from 'react-jsonschema-form'
 
 import { isNotApicastPolicy } from 'Policies/components/util'
 
-import type { ThunkAction } from 'Policies/types/index'
-import type { ChainPolicy } from 'Policies/types/Policies'
+import type { ThunkAction, ChainPolicy } from 'Policies/types'
 import type { UpdatePolicyConfigAction } from 'Policies/actions/PolicyConfig'
 
 class PolicyForm extends Form {
