+++ app/javascript/src/Policies/components/Root.jsx
@@ -2,7 +2,6 @@
 
 import React from 'react'
 import PropTypes from 'prop-types'
-
 import { Provider } from 'react-redux'
 
 import App from 'Policies/components/App'
