+++ app/javascript/src/Policies/reducers/UISettings.jsx
@@ -3,7 +3,7 @@
 import { initialState } from 'Policies/reducers/initialState'
 import { createReducer, updateError, updateObject } from 'Policies/util'
 
-import type { State } from 'Policies/types/State'
+import type { State, UIState } from 'Policies/types'
 import type { UIComponentTransitionAction } from 'Policies/actions/UISettings'
 
 function updateComponentTransition (state: State, action: UIComponentTransitionAction): State {
@@ -16,10 +16,8 @@ function updateRequestsCounter (number: number) {
   }
 }
 
-// eslint-disable-next-line space-infix-ops
-// const UISettingsReducer = createReducer<UIState>(initialState.ui, {
-// $FlowFixMe TODO: in order to fully type createReducer, set UIState and re-enable flow. (use lines above)
-const UISettingsReducer = createReducer(initialState.ui, {
+// TODO: use combineReducers instead of createReducer
+const UISettingsReducer = createReducer<UIState>(initialState.ui, {
   'UI_COMPONENT_TRANSITION': updateComponentTransition,
   'FETCH_CHAIN_ERROR': updateError,
   'FETCH_REGISTRY_ERROR': updateError,
