+++ app/javascript/src/Policies/reducers/index.jsx
@@ -8,7 +8,7 @@ import PolicyConfigReducer from 'Policies/reducers/PolicyConfig'
 import ChainReducer from 'Policies/reducers/PolicyChain'
 import OriginalChainReducer from 'Policies/reducers/OriginalPolicyChain'
 
-// $FlowFixMe: Redux types should work out of the box
+// $FlowFixMe: provide type for State, Action after removing createReducer
 const rootReducer = combineReducers({
   chain: ChainReducer,
   originalChain: OriginalChainReducer,
